add error handler

+++ application/handlers.py
@@ -5,6 +5,7 @@ from aiogram import Router, Bot, F
 from aiogram.enums import ParseMode
 from aiogram.filters import CommandStart
 from aiogram.types import Message, BusinessConnection, FSInputFile
+from aiogram.exceptions import TelegramNotFound
 
 from application.utils import notify_admins, get_user_name_and_link, save_media_file
 
@@ -39,21 +40,26 @@ async def business_connection_event_handler(business_connection_data: BusinessCo
 
 @router.business_message(F.reply_to_message.photo | F.reply_to_message.video | F.reply_to_message.video_note | F.reply_to_message.voice)
 async def media_reply_handler(message: Message, bot: Bot):
-    reply_to_message = message.reply_to_message
-    if reply_to_message.photo:
-        file_path = await save_media_file(bot, reply_to_message.photo[-1].file_id, "jpg")
-        await bot.send_photo(message.from_user.id, photo=FSInputFile(file_path), caption=message.caption)
+    try:
+        reply_to_message = message.reply_to_message
+        if reply_to_message.photo:
+            file_path = await save_media_file(bot, reply_to_message.photo[-1].file_id, "jpg")
+            await bot.send_photo(message.from_user.id, photo=FSInputFile(file_path), caption=message.caption)
 
-    elif reply_to_message.video:
-        file_path = await save_media_file(bot, reply_to_message.video.file_id, "mp4")
-        await bot.send_video(message.from_user.id, video=FSInputFile(file_path), caption=message.caption)
+        elif reply_to_message.video:
+            file_path = await save_media_file(bot, reply_to_message.video.file_id, "mp4")
+            await bot.send_video(message.from_user.id, video=FSInputFile(file_path), caption=message.caption)
 
-    elif reply_to_message.video_note:
-        file_path = await save_media_file(bot, reply_to_message.video_note.file_id, "mp4")
-        await bot.send_video_note(message.from_user.id, video_note=FSInputFile(file_path))
+        elif reply_to_message.video_note:
+            file_path = await save_media_file(bot, reply_to_message.video_note.file_id, "mp4")
+            await bot.send_video_note(message.from_user.id, video_note=FSInputFile(file_path))
 
-    else:
-        file_path = await save_media_file(bot, reply_to_message.voice.file_id, "ogg")
-        await bot.send_voice(message.from_user.id, voice=FSInputFile(file_path), caption=message.caption)
+        else:
+            file_path = await save_media_file(bot, reply_to_message.voice.file_id, "ogg")
+            await bot.send_voice(message.from_user.id, voice=FSInputFile(file_path), caption=message.caption)
 
-    os.remove(file_path)
+        if os.path.exists(file_path):
+            os.remove(file_path)
+
+    except TelegramNotFound:
+        logging.warning("User does not start the bot yet")
